chore(location): move files to a dynamic path
