+++ src/app/(frontend)/(inner)/location/LogoSlider/index.tsx
