+++ src/app/(frontend)/(inner)/location/WorkSlider/index.tsx
